[FIX] hasNext 계산에서 page + 1 오버플로를 막는다
page 에는 상한이 없다(@Min(0) 뿐). Integer.MAX_VALUE 가 들어오면 page + 1 이
Integer.MIN_VALUE 로 돌아 totalPages 보다 작아지고, 빈 페이지인데 hasNext 가
true 로 나갔다.

- page 를 long 으로 올려 비교한다
- Integer.MAX_VALUE 회귀 테스트 추가

CodeRabbit 리뷰 반영 (PR #307).

+++ src/main/java/com/module06/backend/identity/member/application/dto/MemberPage.java
@@ -20,6 +20,7 @@ public record MemberPage(
     public static MemberPage of(long totalElements, int page, int size, List<MemberListItem> content) {
         /* size 는 컨트롤러가 1 이상으로 막지만, 0 이 들어와도 0 나눗셈으로 터지지 않게 한다. */
         int totalPages = size <= 0 ? 0 : (int) ((totalElements + size - 1) / size);
-        return new MemberPage(totalElements, totalPages, page + 1 < totalPages, page, size, content);
+        /* page 에는 상한이 없다 — int 로 더하면 Integer.MAX_VALUE 가 음수로 돌아 hasNext 가 뒤집힌다. */
+        return new MemberPage(totalElements, totalPages, (long) page + 1 < totalPages, page, size, content);
     }
 }
